+++ HISTORY.rst
@@ -1,6 +1,15 @@
 Release History
 ===============
 
+2.1.1 (2015-06-09)
+------------------
+
+* Automatically detect upload port using VID:PID board settings
+  (`issue #231 <https://github.com/platformio/platformio/issues/231>`_)
+* Improved detection of build changes
+* Avoided ``LibInstallDependencyError`` when more then 1 library is found
+  (`issue #229 <https://github.com/platformio/platformio/issues/229>`_)
+
 2.1.0 (2015-06-03)
 ------------------